src/requirements.txt
pyyaml==5.3.1
requests==2.25.1
PyPDF2==1.26.0
fpdf==1.7.2